+++ test/maru_router_test.exs
@@ -62,11 +62,16 @@ defmodule Maru.RouterTest do
           route_param :param do
             def resource, do: @resource
           end
+
+          route_param :param_with_options, type: Integer do
+            def param_context, do: @param_context
+          end
         end
       end
     end
 
     assert %Maru.Router.Resource{path: ["level1", "level2", :param]} = ResourcesTest.resource
+    assert [%Maru.Router.Param{attr_name: :param_with_options, parser: Integer, required: true}] = ResourcesTest.param_context
   end
 
   test "complex resources" do