+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## Changelog
-## v0.9.4-dev
+## v0.9.4 (2016-3-5)
 * Enhancements
   * Allow param type definition in route_param.